chore: change type on duration Circulito animation

+++ lib/src/circulito_animation.dart
@@ -1,9 +1,18 @@
 import 'package:flutter/material.dart';
 
+/// Animation to the drawing of sections.
 class CirculitoAnimation {
-  final double duration;
+  /// The duration of the animation in milliseconds.
+  ///
+  /// Defaults to `500`.
+  final int duration;
+
+  /// The curve of the animation.
+  ///
+  /// Defaults to `Curves.easeInOut`.
   final Curve curve;
 
+  /// Creates an animation to the drawing of sections.
   CirculitoAnimation({
     this.duration = 500,
     this.curve = Curves.easeInOut,
